+++ src/main/java/qna/exception/CannotDeleteException.java
@@ -1,6 +1,7 @@
-package qna;
+package qna.exception;
 
-public class CannotDeleteException extends Exception {
+public class CannotDeleteException extends RuntimeException {
+
     private static final long serialVersionUID = 1L;
 
     public CannotDeleteException(String message) {
