+++ src/main/java/qna/exception/CannotUpdateException.java
@@ -1,0 +1,8 @@
+package qna.exception;
+
+public class CannotUpdateException extends RuntimeException {
+
+    public CannotUpdateException(String message) {
+        super(message);
+    }
+}
