+++ src/main/java/qna/exception/ForbiddenException.java
@@ -1,6 +1,7 @@
-package qna;
+package qna.exception;
 
 public class ForbiddenException extends RuntimeException {
+
     public ForbiddenException() {
     }
