+++ src/main/java/qna/exception/NotFoundException.java
@@ -1,4 +1,5 @@
-package qna;
+package qna.exception;
 
 public class NotFoundException extends RuntimeException {
+
 }
