+++ src/main/java/qna/exception/UnAuthorizedException.java
@@ -1,6 +1,7 @@
-package qna;
+package qna.exception;
 
 public class UnAuthorizedException extends RuntimeException {
+
     private static final long serialVersionUID = 1L;
 
     public UnAuthorizedException() {
@@ -8,7 +9,7 @@ public class UnAuthorizedException extends RuntimeException {
     }
 
     public UnAuthorizedException(String message, Throwable cause, boolean enableSuppression,
-                                 boolean writableStackTrace) {
+            boolean writableStackTrace) {
         super(message, cause, enableSuppression, writableStackTrace);
     }
